preparing for qualifications

+++ MicroMouse/lib/PID/PID.cpp
@@ -1,10 +1,10 @@
 #include "PID.h"
 
-PID::PID(double kp, double ki, double kd, int max_output, int windup_limit)
+PID::PID(float kp, float ki, float kd, int max_output, int windup_limit)
         :kp(kp), ki(ki), kd(kd), max_output(max_output), windup_limit(windup_limit), 
         setpoint(0), prevError(0), integral(0), lastTime(0), direction(0), feedback(0){}
 
-void PID::setSetpoint(double setpoint) {
+void PID::setSetpoint(float setpoint) {
     this->setpoint = setpoint;
 }
 
@@ -16,16 +16,16 @@ int PID::getFeedback() {
     return this->feedback;
 }
 
-double PID::getError(){
+float PID::getError(){
     return setpoint - feedback;
 }
 
-double PID::compute() {
-    double error = setpoint - feedback;
+float PID::compute() {
+    float error = setpoint - feedback;
     prevError = error;
     unsigned long currentTime = millis();
 
-    double deltaTime = (currentTime - lastTime) / 1000.0;
+    float deltaTime = (currentTime - lastTime) / 1000.0;
     lastTime = currentTime;
     integral += error * deltaTime;
 
@@ -33,9 +33,9 @@ double PID::compute() {
         integral = windup_limit;
     }
 
-    double derivative = (error - prevError) / deltaTime;
+    float derivative = (error - prevError) / deltaTime;
 
-    double output = kp * error + ki * integral + kd * derivative;
+    float output = kp * error + ki * integral + kd * derivative;
 
     output > 0 ? this->direction = 1 : this->direction = -1;
